requirements.txt
python-dotenv==1.1.1
zabbix-utils==2.0.3